# Optimize performance of executor when running ops less than number of parallelism