--- a/mars/dataframe/groupby/aggregation.py
+++ b/mars/dataframe/groupby/aggregation.py
@@ -240,23 +240,27 @@ class DataFrameGroupByAgg(DataFrameOperand, DataFrameOperandMixin):
                 elif f == 'count':
                     _add_column_to_functions(col, f, [f], ['sum'], 'sum')
                 elif f == 'mean':
-                    def _mean(df, columns):
-                        return df[columns[0]].sum() / df[columns[1]].sum()
+                    def _mean(_, grouped, columns):
+                        return grouped[columns[0]].sum() / grouped[columns[1]].sum()
 
                     _add_column_to_functions(col, f, ['sum', 'count'], ['sum', 'sum'], _mean)
                 elif f in {'var', 'std'}:
-                    def _reduce_var(df, columns):
-                        cnt = df[columns[1]]
-                        reduced_cnt = cnt.sum()
+                    def _reduce_var(df, grouped, columns):
+                        grouper = grouped.grouper
+
                         data = df[columns[0]]
+                        cnt = df[columns[1]]
                         var_square = df[columns[2]] * (cnt - 1)
-                        avg = data.sum() / reduced_cnt
-                        avg_diff = data / cnt - avg
-                        var_square = var_square.sum() + (cnt * avg_diff ** 2).sum()
+
+                        reduced_cnt = grouped[columns[1]].sum()
+                        avg = grouped[columns[0]].sum() / reduced_cnt
+                        avg_diff_square = (data / cnt - avg.loc[df.index]) ** 2
+                        var_square = var_square.groupby(grouper).sum() + \
+                                     (cnt * avg_diff_square).groupby(grouper).sum()
                         return var_square / (reduced_cnt - 1)
 
-                    def _reduce_std(df, columns):
-                        return np.sqrt(_reduce_var(df, columns))
+                    def _reduce_std(df, grouped, columns):
+                        return np.sqrt(_reduce_var(df, grouped, columns))
 
                     _add_column_to_functions(col, f, ['sum', 'count', 'var'],
                                              ['sum', 'sum', _reduce_var],
@@ -484,27 +488,22 @@ class DataFrameGroupByAgg(DataFrameOperand, DataFrameOperandMixin):
                 # force to get grouped again by copy
                 grouped = cls._get_grouped(op, df, copy=True)
                 result = grouped.agg(func)
-            result.columns = processed_cols
         else:
             # agg the funcs that can be done
-            d = OrderedDict()
-            col_iter = iter(processed_cols)
-            for field, funcs in func.items():
-                for f in funcs:
-                    try:
-                        d[next(col_iter)] = getattr(grouped[field], f)()
-                    except ValueError:
-                        # fail due to buffer read-only
-                        # force to get grouped again by copy
-                        grouped = cls._get_grouped(op, df, copy=True)
-                        d[next(col_iter)] = getattr(grouped[field], f)()
-            result = type(df)(d)
+            try:
+                result = grouped.agg(func)
+            except ValueError:  # pragma: no cover
+                # fail due to buffer read-only
+                # force to get grouped again by copy
+                grouped = cls._get_grouped(op, df, copy=True)
+                result = grouped.agg(func)
+        result.columns = processed_cols
 
         if len(op.output_column_to_func) > 0:
             # process the functions that require operating on the grouped data
             for out_col, f in op.output_column_to_func.items():
                 if len(df) > 0:
-                    result[out_col] = grouped.apply(f)
+                    result[out_col] = f(df, grouped)
                 else:
                     result[out_col] = []