--- a/mars/dataframe/merge/concat.py
+++ b/mars/dataframe/merge/concat.py
@@ -195,31 +195,37 @@ class DataFrameConcat(DataFrameOperand, DataFrameOperandMixin):
         def _auto_concat_dataframe_chunks(chunk, inputs):
             if chunk.op.axis is not None:
                 return pd.concat(inputs, axis=op.axis)
+
             # auto generated concat when executing a DataFrame
-            n_rows = max(inp.index[0] for inp in chunk.inputs) + 1
-            n_cols = int(len(inputs) // n_rows)
-            assert n_rows * n_cols == len(inputs)
-
-            xdf = pd if isinstance(inputs[0], (pd.DataFrame, pd.Series)) else cudf
-
-            concats = []
-            for i in range(n_rows):
-                if n_cols == 1:
-                    concats.append(inputs[i])
-                else:
-                    concat = xdf.concat([inputs[i * n_cols + j] for j in range(n_cols)], axis=1)
-                    concats.append(concat)
-
-            if xdf is pd:
-                # The `sort=False` is to suppress a `FutureWarning` of pandas, when the index or column of chunks to
-                # concatenate is not aligned, which may happens for certain ops.
-                #
-                # See also Note [Columns of Left Join] in test_merge_execution.py.
-                ret = xdf.concat(concats, sort=False)
+            if len(inputs) == 1:
+                ret = inputs[0]
             else:
-                ret = xdf.concat(concats)
-                # cuDF will lost index name when concat two seriess.
-                ret.index.name = concats[0].index.name
+                n_rows = max(inp.index[0] for inp in chunk.inputs) + 1
+                n_cols = int(len(inputs) // n_rows)
+                assert n_rows * n_cols == len(inputs)
+
+                xdf = pd if isinstance(inputs[0], (pd.DataFrame, pd.Series)) else cudf
+
+                concats = []
+                for i in range(n_rows):
+                    if n_cols == 1:
+                        concats.append(inputs[i])
+                    else:
+                        concat = xdf.concat([inputs[i * n_cols + j] for j in range(n_cols)], axis=1)
+                        concats.append(concat)
+
+                if xdf is pd:
+                    # The `sort=False` is to suppress a `FutureWarning` of pandas,
+                    # when the index or column of chunks to concatenate is not aligned,
+                    # which may happens for certain ops.
+                    #
+                    # See also Note [Columns of Left Join] in test_merge_execution.py.
+                    ret = xdf.concat(concats, sort=False)
+                else:
+                    ret = xdf.concat(concats)
+                    # cuDF will lost index name when concat two seriess.
+                    ret.index.name = concats[0].index.name
+
             if getattr(chunk.index_value, 'should_be_monotonic', False):
                 ret.sort_index(inplace=True)
             if getattr(chunk.columns_value, 'should_be_monotonic', False):
@@ -231,19 +237,25 @@ class DataFrameConcat(DataFrameOperand, DataFrameOperandMixin):
             if all(np.isscalar(inp) for inp in inputs):
                 return pd.Series(inputs)
             else:
-                xdf = pd if isinstance(inputs[0], pd.Series) else cudf
-                if chunk.op.axis is not None:
-                    concat = xdf.concat(inputs, axis=chunk.op.axis)
+                if len(inputs) == 1:
+                    concat = inputs[0]
                 else:
-                    concat = xdf.concat(inputs)
+                    xdf = pd if isinstance(inputs[0], pd.Series) else cudf
+                    if chunk.op.axis is not None:
+                        concat = xdf.concat(inputs, axis=chunk.op.axis)
+                    else:
+                        concat = xdf.concat(inputs)
                 if getattr(chunk.index_value, 'should_be_monotonic', False):
                     concat.sort_index(inplace=True)
                 return concat
 
         def _auto_concat_index_chunks(chunk, inputs):
-            xdf = pd if isinstance(inputs[0], pd.Index) else cudf
-            empty_dfs = [xdf.DataFrame(index=inp) for inp in inputs]
-            concat_df = xdf.concat(empty_dfs, axis=0)
+            if len(inputs) == 1:
+                concat_df = inputs[0]
+            else:
+                xdf = pd if isinstance(inputs[0], pd.Index) else cudf
+                empty_dfs = [xdf.DataFrame(index=inp) for inp in inputs]
+                concat_df = xdf.concat(empty_dfs, axis=0)
             if getattr(chunk.index_value, 'should_be_monotonic', False):
                 concat_df.sort_index(inplace=True)
             return concat_df.index
@@ -251,10 +263,7 @@ class DataFrameConcat(DataFrameOperand, DataFrameOperandMixin):
         chunk = op.outputs[0]
         inputs = [ctx[input.key] for input in op.inputs]
 
-        if len(inputs) == 1:
-            # no need to concat
-            ctx[chunk.key] = inputs[0]
-        elif isinstance(inputs[0], tuple):
+        if isinstance(inputs[0], tuple):
             ctx[chunk.key] = tuple(_base_concat(chunk, [input[i] for input in inputs])
                                    for i in range(len(inputs[0])))
         else: