--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -1,5 +1,6 @@
 numpy>=1.14.0
-pandas>=0.24.0
+pandas>=0.24.0; python_version >= '3.6'
+pandas>=0.24.0,<1.0.0; python_version < '3.6'
 requests>=2.4.0
 pyarrow>=0.11.0,<0.16.0
 cython>=0.29