<!--
Thank you for your contribution!

Please review https://github.com/mars-project/mars/blob/master/CONTRIBUTING.rst before opening a pull request.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

When running ops less than the number of parallelism a.k.a `n_parallel`, the performance of executor will become terrible.

Master branch, the file is around 500M, default `chunk_bytes` is 128M, and automatic `n_parallel` is 8.

```
In [1]: %%time 
   ...: import mars.dataframe as md 
   ...: df = md.read_csv('Downloads/ratings.csv') 
   ...: df.groupby('movieId').agg({'rating': ['max', 'min']}).execute() 
   ...:  
   ...:                                                                         
CPU times: user 14.4 s, sys: 2.49 s, total: 16.9 s
Wall time: 8.9 s
Out[1]: 
        rating     
           max  min
movieId            
1          5.0  0.5
2          5.0  0.5
3          5.0  0.5
4          5.0  0.5
5          5.0  0.5
...        ...  ...
131254     4.0  4.0
131256     4.0  4.0
131258     2.5  2.5
131260     3.0  3.0
131262     4.0  4.0

[26744 rows x 2 columns]
```

This PR:

```
In [2]: %%time 
   ...: import mars.dataframe as md 
   ...: df = md.read_csv('Downloads/ratings.csv') 
   ...: df.groupby('movieId').agg({'rating': ['max', 'min']}).execute() 
   ...:  
   ...:                                                                         
CPU times: user 5.53 s, sys: 1.97 s, total: 7.5 s
Wall time: 1.81 s
Out[2]: 
        rating     
           max  min
movieId            
1          5.0  0.5
2          5.0  0.5
3          5.0  0.5
4          5.0  0.5
5          5.0  0.5
...        ...  ...
131254     4.0  4.0
131256     4.0  4.0
131258     2.5  2.5
131260     3.0  3.0
131262     4.0  4.0

[26744 rows x 2 columns]
```

Performance of `groupby.agg` including `mean`, `var` and `std` has been improved quite a lot.

Master:

```
In [2]: %%time 
   ...: df = md.read_csv('Downloads/ratings.csv', chunk_bytes='64M') 
   ...: df.groupby('movieId').agg({'rating': ['mean', 'max', 'min', 'var']}).exe
   ...: cute() 
   ...:  
   ...:                                                                         
/Users/qinxuye/Workspace/mars/mars/dataframe/groupby/aggregation.py:256: RuntimeWarning: invalid value encountered in double_scalars
  return var_square / (reduced_cnt - 1)
CPU times: user 2min 4s, sys: 7.67 s, total: 2min 12s
Wall time: 1min 57s
Out[2]: 
           rating                    
             mean  max  min       var
movieId                              
1        3.921240  5.0  0.5  0.790342
2        3.211977  5.0  0.5  0.904686
3        3.151040  5.0  0.5  1.013328
4        2.861393  5.0  0.5  1.200563
5        3.064592  5.0  0.5  0.964598
...           ...  ...  ...       ...
131254   4.000000  4.0  4.0       NaN
131256   4.000000  4.0  4.0       NaN
131258   2.500000  2.5  2.5       NaN
131260   3.000000  3.0  3.0       NaN
131262   4.000000  4.0  4.0       NaN

[26744 rows x 4 columns]
```

This PR:

```
In [3]: %%time 
   ...: df = md.read_csv('Downloads/ratings.csv', chunk_bytes='64M') 
   ...: df.groupby('movieId').agg({'rating': ['mean', 'max', 'min', 'var']}).exe
   ...: cute() 
   ...:  
   ...:                                                                         
CPU times: user 9.62 s, sys: 3.46 s, total: 13.1 s
Wall time: 1.86 s
Out[3]: 
           rating                    
             mean  max  min       var
movieId                              
1        3.921240  5.0  0.5  0.790342
2        3.211977  5.0  0.5  0.904686
3        3.151040  5.0  0.5  1.013328
4        2.861393  5.0  0.5  1.200563
5        3.064592  5.0  0.5  0.964598
...           ...  ...  ...       ...
131254   4.000000  4.0  4.0       NaN
131256   4.000000  4.0  4.0       NaN
131258   2.500000  2.5  2.5       NaN
131260   3.000000  3.0  3.0       NaN
131262   4.000000  4.0  4.0       NaN

[26744 rows x 4 columns]
```

Tasks include:

- [x] Optimize executor.
- [x] Optimize `mean`, `var` and `std` for `groupby.agg`.

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this change? -->

NA